requirements.txt
tensorflow==2.13.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
requests==2.31.0
tqdm==4.65.0
PyYAML==6.0
matplotlib==3.7.1
seaborn==0.12.2
